create group working

+++ ynetwork/database/migrations/2023_10_29_212944_create_post.php
@@ -11,23 +11,24 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('post', function (Blueprint $table) {
-            $table->id();
-            $table->integer('likes');
-            $table->integer('dislikes');
-            $table->string('content');
-            $table->string('subject');
-            $table->unsignedInteger('reactsTo')->nullable(); // Změněno na 'unsignedInteger' pro foreign klíč
-            $table->unsignedInteger('groupid'); // Změněno na 'unsignedInteger' pro foreign klíč
-            $table->unsignedInteger('userid'); // Změněno na 'unsignedInteger' pro foreign klíč
-            $table->boolean('isVerified');
-        });
-
-        Schema::table('post', function (Blueprint $table) {
+        if(!Schema::hasTable('post')){
+            Schema::create('post', function (Blueprint $table) {
+                $table->id();
+                $table->integer('likes');
+                $table->integer('dislikes');
+                $table->string('content');
+                $table->string('subject');
+                $table->unsignedInteger('reactsTo')->nullable(); // Změněno na 'unsignedInteger' pro foreign klíč
+                $table->unsignedInteger('groupid'); // Změněno na 'unsignedInteger' pro foreign klíč
+                $table->unsignedInteger('userid'); // Změněno na 'unsignedInteger' pro foreign klíč
+                $table->boolean('isVerified');
+            });
+        }
+        /* Schema::table('post', function (Blueprint $table) {
             $table->foreign('reactsTo')->references('id')->on('post'); // Nastavení foreign klíče
             $table->foreign('groupid')->references('id')->on('group'); // Nastavení foreign klíče
             $table->foreign('userid')->references('id')->on('user'); // Nastavení foreign klíče
-        });
+        }); */
     }
 
     /**
@@ -35,6 +36,8 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('post');
+        if (Schema::hasTable('post')) {
+            Schema::dropIfExists('post');
+        }
     }
 };
